+++ janitor-orm/src/main/java/com/eischet/janitor/orm/dao/Dao.java
@@ -43,12 +43,14 @@ public interface Dao<T extends OrmEntity> extends JanitorObject {
     List<T> findByQuery(@NotNull DatabaseConnection conn, @NotNull @Language("SQL") String query, @NotNull StatementConfigurator statementConfigurator) throws DatabaseError;
 
 
+    @Deprecated // use findByFilter(DatabaseConnection, FilterQuery) instead
     @NotNull
     @Unmodifiable
     List<T> findByFilter(@NotNull DatabaseConnection conn,
                          @NotNull FilterExpression filterExpression,
                          @Nullable Integer limit) throws DatabaseError;
 
+    @Deprecated // use findByFilter(DatabaseConnection, FilterQuery) instead
     @NotNull
     @Unmodifiable
     List<T> findByFilter(@NotNull DatabaseConnection conn,
@@ -57,12 +59,15 @@ public interface Dao<T extends OrmEntity> extends JanitorObject {
                          @Nullable Integer limit) throws DatabaseError;
 
 
+    @Deprecated // use findByFilter(DatabaseConnection, FilterQuery) instead
     @NotNull
     @Unmodifiable
     default List<T> findByFilter(@NotNull DatabaseConnection conn,
                                  @NotNull FilterExpression filterExpression) throws DatabaseError {
         return findByFilter(conn, filterExpression, null);
     }
+
+    @NotNull @Unmodifiable List<T> findByFilter(@NotNull DatabaseConnection conn, @NotNull FilterQuery filterQuery) throws DatabaseError;
 
     int countByFilter(@NotNull DatabaseConnection conn, @Nullable FilterExpression filterExpression) throws DatabaseError;